# Dhruv_Chavda_150096725060_Library_Management/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Library Management System API
  description: >-
    A complete RESTful API for a Library Management System built with Node.js,
    Express, JWT authentication, and Google Firebase Firestore. Supports
    role-based access control for Students and Librarians.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token in the format **Bearer <token>**
  schemas:
...
    User:
      type: object
      properties:
        userId:
          type: string
          example: usr_872364
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: student@example.com
        role:
          type: string
          enum: [student, librarian]
          example: student
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Book:
      type: object
      properties:
        bookId:
          type: string
          example: bk_982341
        title:
          type: string
          example: Clean Code
        author:
          type: string
          example: Robert C. Martin
        isbn:
          type: string
          example: 978-0132350884
        category:
          type: string
          example: Software Engineering
        status:
          type: string
          enum: [available, borrowed]
          example: available
        quantity:
          type: integer
          example: 5
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BookCreateRequest:
      type: object
      required:
        - title
        - author
        - isbn
        - category
        - quantity
      properties:
        title:
          type: string
          example: Clean Code
        author:
          type: string
          example: Robert C. Martin
        isbn:
          type: string
          example: 978-0132350884
        category:
          type: string
          example: Software Engineering
        quantity:
          type: integer
          minimum: 1
          example: 5
...
    Transaction:
      type: object
      properties:
        transactionId:
          type: string
          example: tx_102938
        userId:
          type: string
          example: usr_872364
        bookId:
          type: string
          example: bk_982341
        type:
          type: string
          enum: [borrow, return]
          example: borrow
        borrowDate:
          type: string
          format: date-time
        returnDate:
          type: string
          format: date-time
          nullable: true
        dueDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, returned, overdue]
          example: active
        createdAt:
          type: string
          format: date-time
paths:
...
  /api/books:
    get:
      summary: Get all books (with optional filters)
      tags: [Books]
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: Filter books by category
        - in: query
          name: status
          schema:
            type: string
            enum: [available, borrowed]
          description: Filter books by availability status
        - in: query
          name: author
          schema:
            type: string
          description: Filter books by author name
      responses:
        200:
          description: List of books
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 10
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
    post:
      summary: Add new book (Librarian only)
      tags: [Books]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookCreateRequest'
      responses:
        201:
          description: Book added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Book'
        403:
          description: Forbidden (Librarian only)
...
  /api/transactions:
    get:
      summary: Get all transactions across library (Librarian only)
      tags: [Transactions]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
        403:
          description: Forbidden (Librarian only)
...
  /api/users:
    get:
      summary: Get all users (Librarian only)
      tags: [User Management]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of all users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        403:
          description: Forbidden (Librarian only)
...
  /api/test-firebase:
    get:
      summary: Test Firebase Firestore database connection
      tags: [System]
      responses:
        200:
          description: Firebase connected successfully
        500:
          description: Firebase connection error